Rename parse/symbol -> parse/parsers
