+++ src/parse/parsers/statement/declaration.rs
@@ -6,7 +6,7 @@
 use lex::{Token, Tokenizer, TokenType};
 use ast::*;
 use parse::{Parser, ParseResult};
-use parse::symbol::{PrefixParser, Precedence};
+use parse::parsers::{PrefixParser, Precedence};
 
 ///
 /// # Examples
