+++ src/parse/parsers/statement/do_block.rs
@@ -3,7 +3,7 @@
 use lex::{Token, Tokenizer, TokenType};
 use ast::*;
 use parse::{Parser, ParseResult};
-use parse::symbol::PrefixParser;
+use parse::parsers::PrefixParser;
 
 /// Parses a block statement using the prefix symol `do`.
 ///
